requirements.txt
flask 
pdfplumber
python-docx
pdfminer.six